requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
pymongo==4.4.1

# Environment
python-dotenv==1.0.0

# Cloud LLM
groq==0.4.1

# Hugging Face Gradio API
gradio_client>=1.13.0

# Hugging Face Models
transformers==4.35.0
huggingface_hub==0.18.0

# Embeddings & RAG
sentence-transformers==2.2.2

# Deep Learning (Image Classification)
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0

# Data Processing
numpy==1.24.3
scikit-learn==1.3.0

# HTTP Requests
requests==2.31.0

# Audio (Optional)
SpeechRecognition==3.10.0
pydub==0.25.1
